弾性体クラス Elastic にダンパを追加． GUIサンプルに Hair　を追加． 検証のために，外力設定とレンダリング設定が必要．

+++ trunk/Crystal/CrystalRenderer/PointSprite.h
@@ -38,7 +38,7 @@ public:
 
 	const FrameBuffer& getBuffer() const { return buffer; }
 
-	const int getRadius() const { return buffer.getWidth() / 2; }
+	const size_t getRadius() const { return buffer.getWidth() / 2; }
 
 private:
 	FrameBuffer buffer;
